+++ parserShoeBoxAtlas/parserShoeBoxAtlas.cpp
@@ -236,12 +236,12 @@ bool Atlas::is_open(){
 }
 
 // versão de drawCropAtlas com opt igual a 0
-void Atlas::drawCropAtlas(int posX, int posY){
+void Atlas::drawCropAtlas(float posX, float posY){
     drawCropAtlas(posX, posY, 0);
 }
 
 // desenha parte do atlas em um bitmap
-void Atlas::drawCropAtlas(int posX, int posY, int opt){
+void Atlas::drawCropAtlas(float posX, float posY, int opt){
     al_draw_bitmap_region(atlas, currentSprite.posX, currentSprite.posY,
                              currentSprite.w, currentSprite.h, currentSprite.offX+posX,
                              currentSprite.offY+posY,opt);
